test: fold the new Unstructured unit tests into test_converter.py
Address review feedback on #3794:

- Merge test_converter_unit.py into test_converter.py so the converter's
  tests live in one file, with the api_key=None serialization round-trip
  next to the existing to_dict/from_dict tests. Reuse the new LOCAL_API_URL
  constant in the existing integration tests instead of repeating the
  literal ten times.
- Pin the ValueError message in the metadata-length-mismatch test, since
  run() has two ValueError paths.
- Put a file inside the subdirectory in the directory test, so it actually
  pins the non-recursive glob rather than only the is_file() guard.
